src/cmakeliste a lot cleaner already, now need to make it global and available for apps and test. Moved transformpayload to parser folder.
