--- src/data/bsfdatabaseconfig.h
@@ -1,23 +1,0 @@
-#ifndef BSF_DATABASECONFIG_H
-#define BSF_DATABASECONFIG_H
-
-#include <QtCore/QString>
-#include <QtSql/QSqlDatabase>
-
-class BsfDbconfig {
-
-public:
-    BsfDbconfig();
-    const QString defaultDatabaseName = "data/bsf.db";
-    const QString database = "QSQLITE";
-    const QString defaultConnection = "qt_sql_default_connection";
-    void initDatabaseConnection();
-
-    void setDatabaseName(const QString &);
-    void setSqlDatabase(QSqlDatabase&);
-
-private:
-    QString databaseName = "";
-
-};
-#endif //BSF_DATABASECONFIG_H
