+++ src/service/iodeviceservice.cpp
@@ -1,4 +1,4 @@
-#include "service/iodeviceservice.h"
+#include "iodeviceservice.h"
 
 #include <repo/arduinorepo.h>
 #include <repo/actionarduinorepo.h>
