+++ src/statemachine/bsfpavementmachine.cpp
@@ -1,4 +1,4 @@
-#include "bsfpavementmachine.h"
+#include <statemachine/bsfpavementmachine.h>
 
 void BsfPavementMachine::setPavementRecipe(RecipeInfoData *pData) {
     BEGIN_TRANSITION_MAP                      // - Current State -
