+++ src/ui/main/mainwindow.h
@@ -7,6 +7,7 @@
 #include <QToolBar>
 #include <QStatusBar>
 
+
 class QAction;
 class QMenu;
